==> scripts/1_preprocesamiento_datos_kobo.Rmd <==
---
title: "Preprocesamiento de datos kobo"
output:
  html_document:
    df_print: paged
---

```{r, message=FALSE}
library(dplyr)
library(tidyr)
library(readr)
library(stringr)
```

Data exported from kobo need to be processed to a tidy format that R can manipulate more easily. 

## Trees sampling

Trees sampling was done by the 12 brigadistas in 48 plots of 10 x 10 m, using a kobo form. 

### Load data and check for basic errors

```{r}
# load data
muestreo<-read.delim("../data/kobo/muestreo_dic2020_raw.txt", header = TRUE) %>%
# filter rows w/o actual data
  dplyr::filter(., !is.na(plot)) %>%

# remove unnecesary columns
  select(., -coordinates, -X__version__, -X_version_, -X_id, -X_uuid, -(X_index:X.6))
head(muestreo)
```

Create sample ID concateneting brigadista id + tree number (which were separate questions in kobo):

```{r}
muestreo$tree_id<-paste0(muestreo$person_key, muestreo$tree_number)
```

Check how many trees were sampled:

```{r}
# total
nrow(muestreo)

# dead / alive
table(muestreo$life_condition)

# by brigadista
table(muestreo$person_key)

# by plot
table(muestreo$plot)
```

We only sampled 48 plots, any larger number must be a typo during sampling. Check, and if there is an error, ommit those samples:

```{r}
# check
filter(muestreo, plot > 48) %>% select(plot, person_key, life_condition, tree_id)

# filter from data
muestreo <- filter(muestreo, !plot >48)
```
### Examine and clean tree health data

The variables `tree_health.*` are the result of a multiple choice question according to the tree health state. A single tree can have more than one type of damage.

```{r}
head(select(muestreo, tree_health.healthy:tree_health.other))
```

Change 0 - 1 notation for condition name:

```{r}
# 1 for healthy 
muestreo$tree_health.healthy= gsub("1", "healthy", as.character(muestreo$tree_health.healthy))
muestreo$tree_health.ozone= gsub("1", "ozone", as.character(muestreo$tree_health.ozone))
muestreo$tree_health.fungi= gsub("1", "fungi", as.character(muestreo$tree_health.fungi))
muestreo$tree_health.drougth= gsub("1", "drougth", as.character(muestreo$tree_health.drougth))
muestreo$tree_health.worm= gsub("1", "worm", as.character(muestreo$tree_health.worm))
muestreo$tree_health.insect= gsub("1", "insect", as.character(muestreo$tree_health.insect))
muestreo$tree_health.acid_rain= gsub("1", "acid_rain", as.character(muestreo$tree_health.acid_rain))
muestreo$tree_health.other= gsub("1", "other", as.character(muestreo$tree_health.other))

# 0 for NA
muestreo$tree_health.healthy= gsub("0", NA, as.character(muestreo$tree_health.healthy))
muestreo$tree_health.ozone= gsub("0", NA, as.character(muestreo$tree_health.ozone))
muestreo$tree_health.fungi= gsub("0", NA, as.character(muestreo$tree_health.fungi))
muestreo$tree_health.drougth= gsub("0", NA, as.character(muestreo$tree_health.drougth))
muestreo$tree_health.worm= gsub("0", NA, as.character(muestreo$tree_health.worm))
muestreo$tree_health.insect= gsub("0", NA, as.character(muestreo$tree_health.insect))
muestreo$tree_health.acid_rain= gsub("0", NA, as.character(muestreo$tree_health.acid_rain))
muestreo$tree_health.other= gsub("0", NA, as.character(muestreo$tree_health.other))

# check
head(select(muestreo, tree_health.healthy:tree_health.other))
```

The original `tree_health` variable concatenates kobo answers, but it changes depending in the order the options were selected, so lets create a new variable concatenanting from the binary `tree_heath.*` variables

```{r}
# use useful fuction found in https://stackoverflow.com/questions/13673894/suppress-nas-in-paste
paste5 <- function(..., sep = " ", collapse = NULL, na.rm = F) {
  if (na.rm == F)
    paste(..., sep = sep, collapse = collapse)
  else
    if (na.rm == T) {
      paste.na <- function(x, sep) {
        x <- gsub("^\\s+|\\s+$", "", x)
        ret <- paste(na.omit(x), collapse = sep)
        is.na(ret) <- ret == ""
        return(ret)
      }
      df <- data.frame(..., stringsAsFactors = F)
      ret <- apply(df, 1, FUN = function(x) paste.na(x, sep))

      if (is.null(collapse))
        ret
      else {
        paste.na(ret, sep = collapse)
      }
    }
}

# paste to concatenate in new variable
muestreo$tree_health <- paste5(muestreo$tree_health.healthy, 
                        muestreo$tree_health.ozone, 
                        muestreo$tree_health.fungi,
                        muestreo$tree_health.drougth,
                        muestreo$tree_health.worm,
                        muestreo$tree_health.insect,
                        muestreo$tree_health.acid_rain,
                        muestreo$tree_health.other,
                        na.rm=TRUE)
# check
head(muestreo$tree_health)
```

Check combinations of damage:

```{r}
unique(muestreo$tree_health)
```

We can't have "healthy + a damage condition". If this appears this is an error made during sampling. If this happens we have to check the photos of this entries to correct accordingly. All values of "healthy + a damage" will be ommited from the data.


```{r}
# check if there are any trees with healthy + other. If this is the case, there should be an " " after healthy. Only non-approved records should appear here:
muestreo[grep("healthy ", muestreo$tree_health), c("plot", "tree_id", "X_validation_status")]

# remove any tree labeled with healthy + something else, independently of its validation status
muestreo<-muestreo[grep("healthy ", muestreo$tree_health, invert=TRUE), ]

# check
unique(muestreo$tree_health)
```

Make a new variable to collapse the interaction of ozone + others in a single variable. This would be useful for visualization purposes:

```{r}
# collapse together interactions of ozone + others
muestreo <- muestreo %>%
            mutate(., tree_health_simplified = ifelse(grepl("ozone ", tree_health), "ozone_and_other", tree_health)) %>% 

# collapse together other types of damage
            mutate(., tree_health_simplified = ifelse(grepl(" ", tree_health_simplified), "others_combined", tree_health_simplified))

#check
head(muestreo$tree_health_simplified)
table(muestreo$tree_health_simplified)
```

Samples with `NA` in `tree_health_simplified` are actually dead trees:

```{r}
filter(muestreo, is.na(tree_health_simplified)) %>% select(life_condition)
```

Change "NA" to "dead" in tree_health_simplified`:

```{r}
muestreo<- mutate(muestreo, 
                  tree_health_simplified=  replace_na(tree_health_simplified, "dead"))
table(muestreo$tree_health_simplified)
```

Columns with presence/absence health condition data are no longer needed, remove them
```{r}
muestreo<- select(muestreo, 
                  -c(tree_health.healthy:tree_health.other))
```

#### Clean percentage of ozone damage by tree

The variables `ozone.damage.percentage*`are the result of a multiple choice question according to the percentage of the tree showing ozone damage (categorical variable). 

See data:
```{r}
head(select(muestreo, ozone_damage_percentage:ozone_damage_percentage.70.100))
```

Change 0 - 1 notation for condition name:

```{r}
# 0 for NA
muestreo$ozone_damage_percentage.less_10 = gsub("0", NA, as.character(muestreo$ozone_damage_percentage.less_10))
muestreo$ozone_damage_percentage.10.40 = gsub("0", NA, as.character(muestreo$ozone_damage_percentage.10.40))
muestreo$ozone_damage_percentage.50 = gsub("0", NA, as.character(muestreo$ozone_damage_percentage.50))
muestreo$ozone_damage_percentage.50.70 = gsub("0", NA, as.character(muestreo$ozone_damage_percentage.50.70))
muestreo$ozone_damage_percentage.70.100 = gsub("0", NA, as.character(muestreo$ozone_damage_percentage.70.100))

#1 for condition
muestreo$ozone_damage_percentage.less_10 = gsub("1", "less than 10%", as.character(muestreo$ozone_damage_percentage.less_10))
muestreo$ozone_damage_percentage.10.40 = gsub("1", "10 to 40%", as.character(muestreo$ozone_damage_percentage.10.40))
muestreo$ozone_damage_percentage.50 = gsub("1", "40 to 50%", as.character(muestreo$ozone_damage_percentage.50))
muestreo$ozone_damage_percentage.50.70 = gsub("1", "50 to 70%", as.character(muestreo$ozone_damage_percentage.50.70))
muestreo$ozone_damage_percentage.70.100 = gsub("1", "more than 70%", as.character(muestreo$ozone_damage_percentage.70.100))

# check
head(select(muestreo, ozone_damage_percentage:ozone_damage_percentage.70.100))

```

The original `ozone_damage_percentage` variable concatenates kobo answers, but it changes depending in the order the options were selected, so lets create a new variable concatenanting from the ozone_damage_percentage.* variables:

```{r}
# paste to concatenate in new variable
muestreo$ozone_damage_percentage<-paste5(muestreo$ozone_damage_percentage.less_10, 
                       muestreo$ozone_damage_percentage.10.40, 
                       muestreo$ozone_damage_percentage.50,
                       muestreo$ozone_damage_percentage.50.70,
                       muestreo$ozone_damage_percentage.70.100,
                       na.rm=TRUE)
# check
head(muestreo$ozone_damage_percentage)
```

Check combinations of damage:

```{r}
unique(muestreo$ozone_damage_percentage)
```

A single tree should have only one category of damage, if this is not the case, it is a mistake that needs to be corrected. Further versions of the kobo form would only allow one option to be selected.

```{r}
# check if there are any trees with more than one damage category
muestreo[grep("% ", muestreo$ozone_damage_percentage), c("plot", "tree_id", "ozone_damage_percentage")]

# remove any tree labeled with more than one damage category
muestreo<-muestreo[grep("% ", muestreo$ozone_damage_percentage, invert=TRUE), ]

# check
unique(muestreo$tree_health)

```

Columns with presence/absence health condition data are no longer needed, remove them

```{r}
muestreo<- select(muestreo, 
                  -c(ozone_damage_percentage.less_10:ozone_damage_percentage.option_6))
```



### Examine and clean reforestation notes

If a reforested plant had a tag, we asked brigadistas to write what the tag said, or to add "NP" if it was not present. Somethings the tag is only the year, but it can also include other information, like "SRX" if it was from Santa Rosa Xochiac.

```{r}
table(muestreo$reforestation_code)
```

First, we change the one uncertain (eg. "NP al parecer 2014") and no data (which means the plant was not reforested and this questions was not answered by brigadistas) to NA: 

```{r}
# as charcter
muestreo$reforestation_code<-as.character(muestreo$reforestation_code)

# change uncertian registry to NP
muestreo$reforestation_code<- gsub("NP al parecer 2014", "NP", muestreo$reforestation_code)
muestreo$reforestation_code<- gsub("NP Distribucion simetrica Sin cajete", "NP", muestreo$reforestation_code)
muestreo$reforestation_code<- gsub("NP. Se nota sin mantenimiento", "NP", muestreo$reforestation_code)

# change "" to NA
muestreo$reforestation_code[muestreo$reforestation_code==""] <- NA

```

Create a new variable with only the year data (keeping NP):

```{r}
# delete SRX
muestreo$reforestation_year<-gsub("SRX", "", muestreo$reforestation_code)
# delete white spaces
muestreo$reforestation_year<-str_trim(muestreo$reforestation_year, side="both")
# check
table(muestreo$reforestation_year)
```

### Check Validation status

The previous code checked for capture errors and re-formated the data accordingly. The validation of each record in terms of the type damage meeting the picture was done manually in the web app. Downstream analyses filter data in order to only perform analyses with approved records. The distribution of approved and not approved records is as follows:

```{r}
table(muestreo$X_validation_status)
```



## Sampling plots

Plots metadata was surveyed by Alicia Mastretta using another kobo form.

#### Load data and check

```{r}
parcelas<-read.delim("../data/kobo/parcelas_dic2020_raw.txt", header=TRUE) %>%

# remove unnecesary columns
  select(., -coordinates, -X__version__, -X_version_, -X_id, -X_uuid, -X_index)
parcelas
```

#### Clean dominant spp

To be done.

#### Add summary of tree health by plot

First make a summary of tree health from the sampling form:

```{r}
#Make muestreo within validation aproved
muestreo2<- filter(muestreo, X_validation_status == "validation_status_approved")


#Make summary 
plots_health <- group_by(muestreo2, plot) %>% 
  summarise(tree_health_simplified) %>%
  
# Pivot wider 
              pivot_wider(., names_from = tree_health_simplified,
                            values_from = tree_health_simplified,
                            values_fn = list(tree_health_simplified = length)) %>%
                  
# replace na with 0 
                replace(is.na(.), 0)

plots_health
```

Then add this summary to the parcelas form:

```{r}
# join tables
parcelas<-inner_join(parcelas, plots_health, by = "plot")
```

### Examine and clean reforestation notes

The origin of the reforestation was asked at the parcelas level asking the brigadistas that were present during each reforestation, but that were not necesarely sampling trees in that parcel.

```{r}
levels(parcelas$reforestation_notes)
```

Create a new variable where only :

```{r}
parcelas$reforestation_notes[parcelas$reforestation_notes=="2013 santa rosa "]<- "2013 vivero santa rosa"
```

Create new variable with only origin
```{r}
gsub("[^[:digit:]]+", " ", parcelas$reforestation_notes)
```



## Save tidy data

Remove unused columns

```{r}
# from parcelas
parcelas<- select(parcelas, 
                  -c(dominant_spp.Pinus_ayacauite:dominant_spp_others),
                  -X_submission_time)
```


Save data
```{r}
write.table(muestreo, file="../data/kobo/muestreo_dic2020_tidy.txt", sep ="\t")
write.table(parcelas, file="../data/kobo/parcelas_dic2020_tidy.txt", sep ="\t")
```
